pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Nikolai Tenengolts <ntenengolts@gmail.com>"]
readme = "README.md"

packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.10"

pyyaml = "^6.0.2"
[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
isort = "^5.13.2"

pytest-cov = "^6.0.0"
[tool.poetry.dev-dependencies]
pytest = "^7.0"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
